requirements.txt
pandas
numpy
scikit-learn
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.28.0
Pillow
tqdm
kagglehub
